+++ bilrost-derive/src/lib.rs
@@ -1,4 +1,4 @@
-#![doc(html_root_url = "https://docs.rs/bilrost-derive/0.1002.0")]
+#![doc(html_root_url = "https://docs.rs/bilrost-derive/0.1002.1")]
 // The `quote!` macro requires deep recursion.
 #![recursion_limit = "4096"]